locking/seqlock, headers: Untangle the spaghetti monster (bsc#1176564 bsc#1162702).
The patches.suse/locking-seqlock-headers-Untangle-the-spaghetti-monst.patch
is amended with additional include fixups which are needed for the
current version of the tree. See the commit message of the patch
for a detailed changelog. Also added the devel-rt upstream fix to fight back
the attack of the include spaghetti monster.

Refresh
- patches.rt/mm-rt-kmap-atomic-scheduling.patch

Added
- patches.rt/mutex-move-the-ww_mutext-definition-back-to-ww_mutex.h.patch

suse-commit: 8bcfa017074204054b72f7ca9512416acf1d9b1a